fix entrance
